Fixed the bug that occurs when no substituions are deemed applicable to users preferences
